server/requirements.txt
asgiref>=3.8.1
Django>=4.2.16
sqlparse>=0.5.1
tzdata>=2024.2
djangorestframework>=3.12,<4.0
django-cors-headers>=4.6.0
matplotlib>=3.9.2
networkx>=3.4.2
python-dotenv>=1.0.1
cryptography>=44.0.0
pycryptodome>=3.21.0
python-dotenv>=1.0.1